Revert "gitk: Only restore window size from ~/.gitk, not position"
This reverts commit b9bee11526ec23541ddbbd75716bfd1acc241463.

The commit catered to an unsupportable port of the Windows Tcl/Tk
stuck at 8.4.1 that was used by Cygwin. 8.4.1 has some bad bugs in its
layout engine, and forced changes in Gitk to be compatible. All this
became irrelevant around 2011 after Cygwin gained an X11 server and
switched to a supportable port of the Unix/X11 Tcl/Tk (it is now on the
current 8.6 code base).

Helped-by: Mark Levedahl <mlevedahl@gmail.com>
Signed-off-by: Johannes Sixt <j6t@kdbg.org>