gitk: Make TclTk 8.6 the minimum, allow 8.7
gitk runs under wish, so naturally has Tcl and Tk available, and of the
same version. gitk sets a requirement on Tk version >= 8.4, but this is
very outdated, and not shipping on any current OS.  As 8.7 is in alpha
test, and is generally compatible with 8.x, we should allow that, but
9.0 has planned compatibility breaking changes so is not yet supported.

Also, as git-gui requires at least 8.5, and gitk and git-gui can launch
each other, the minimum requirement should be at least 8.5.  But, no
currently supported OS is shipping 8.5, much less 8.4, and 9.0 is
already released.  So, even 8.5 is essentially an untestable
requirement: 8.6 is a more reasonable minimum requirement.

So, we should allow 8.6-8.7 but not 9.0. As we are using wish, we only
need to check the version of Tcl or of Tk as these must be the same.
Let's put this at the top of file so the requirements are obvious.

Signed-off-by: Mark Levedahl <mlevedahl@gmail.com>