gitk: windows: make "Auto-select" copy to clipboard
The config option "Auto-select SHA1 (length)" (widget-text-)selects
the first $autosellen chars of the commit hash. However, selection on
its own is arguably not very useful, and also not necessarily visible
when not focused, e.g. with Windows native theme (the default, but is
visible on Windows with other themes, like "clam").

The real goal here to copy the text to the [selection?] clipboard,
which works on X11 where "selection" is synonymous with "clipboard
selection", but not necessariy elsewhere, for instance on Windows
(where selecting the text doesn't copy it), and maybe elsewhere too.

This commit adds explicit copy to clipboard on Windows, which happens
whenever the user clicks a commit-graph entry, and like X11, affected
by the configured length and enable/disable checkbox.

Possible future modifications:
- Fix other platforms if needed, e.g. maybe non-X11 macOS Tk, Wayland.
- Rename Auto-select to Auto-copy (at least at the config dialog).
- Rename SHA1 to "commit hash" (I think git plans a move to SHA256).
- Ensure "foo bar" Auto-select length is rejected at the dialog
  (probably also at other numeric input fields).

Signed-off-by: Avi Halachmi (:avih) <avihpit@yahoo.com>