gitk: add support for Auto-copy (primary clipboard)
Auto-select is useful - auto-copy the commit hash of the selected
commit to the selection clipboard, as frequently the next thing after
using gitk is using the hash with a CLI command, and it can be useful
when it's already at the selection clipboard.

However, Auto-select works only in X11, and until now there wasn't
a similar mechanism on other platforms in gitk.

Add a new Auto-copy preference and config checkbox, which copies the
[short] commit hash to the (primary) clippboar, and which uses the
same configured hash-length value which is used by Auto-select.

However, while Auto-select is enabled by default, Auto-copy isn't.

That's because the selection clipboard is typically dispensable, while
the (primary) clipboard can be considered a more precious resource,
which we don't want to overwrite by default.

Possible future modifications:
- Rename "SHA1" to "hash" or "commit hash" throughout the UI, as it
  might not remain SHA1 forever.
- Ensure "foo bar" Auto-select length is rejected at the dialog, and
  probably the same also at other numeric input fields.

Signed-off-by: Avi Halachmi (:avih) <avihpit@yahoo.com>