gitk: Interpret file names given as command argument as UTF-8
When a command line argument is a file name that is not ASCII, then it
is interpreted as a sequence of bytes when Gitk is invoked in this form:

   gitk あいう.txt

This happens because in this case we invoke `git rev-parse` to filter out
the file names from the command line arguments. But we do so without
asking Tcl to interpret the git's out as UTF-8 (like we do everywhere
else where file names are read from git output), and so we receive just
bytes and interpret them in the current locale. This is a problem on
Windows where the current locale is typically not UTF-8, but file names
are interpreted as as UTF-8 by convention.

Replace the exec command by the usual open-fconfigure-read-close dance.

Note that the problem does not occur when Gitk is invoked with
double-dashes, because then the arguments need not be filtered and the
file names are take from argv directly, which have the correct encoding.

Helped-by: Avi Halachmi (:avih) <avihpit@yahoo.com>
Signed-off-by: Johannes Sixt <j6t@kdbg.org>