use pbcopy on MacOS for better clipboard manager support
On Mac OS, current versions of Tk 9 or Tk 8 `clipboard append` do not
interact properly with clipboard managers. This makes `gitk` interact
badly with them as well. See [1] for an example of how this affects
`gitk` and [2] for a ticket describing the problem with Tk.

This commit works around the problem by having `gitk` use the `pbcopy`
CLI to copy to the clipboard instead of Tk's `clipboard append`, as long
as we are on Mac OS and `pbcopy` is available. AFAIK, `pbcopy` is always
installed with Mac OS.

[1]: https://github.com/p0deje/Maccy/issues/1100
[2]: https://core.tcl-lang.org/tk/tktview/e94c8bc845b4d5a658cd5421ef23ef2484560fa8