gitk: use $config_variables as list for save/restore
gitk offers many variables for configuration, and has all of these
listed in $config_variables. gitk allows modification of many of these
from a dialog box, and this dialog can be cancelled requiring that any
configuration variable modified is restored to its prior value. But,
gitk maintains lists of variables that can be used in the configuration
dialog separately from $config_variables, and these lists exist
independently in more than one place in the code.

This is confusing, error prone, and unnecessary. Let's just use
$config_variables as the authoritative list in all places: a new
configuration variable need only be added to this list to be saved,
restored, and returned to its prior value after the configuration dialog
is cancelled. This does mean that more unchanged global variables are
overwritten with their current values after cancelling the configuration
dialog, but this no-op takes trivial cpu time.

Signed-off-by: Mark Levedahl <mlevedahl@gmail.com>