gitk: set minimum git version at 2.36
gitk checks in a few places for very early git versions, such as 1.72,
but places no restriction on what version is used. gitk is distributed
as part of the full git tree, packagers do not mix parts from different
release versions, so it is unclear what backwards compatibility is
reasonably expected.  While tolerance for a non-current git version is
good for developers, we shouldn't expect the code to add features for
newer git but operate perpetually with git circa 2007.  Also, git-gui
and gitk can both launch each other, these two programs should have
compatible minimum requirements to avoid a confusing or broken
installation.

git-gui needs git-hook-run, added in git 2.36 while git 2.50 is now
current, so 2.36 is a reasonable minimum for now. Make it so.

Signed-off-by: Mark Levedahl <mlevedahl@gmail.com>