gitk: replace ${NS} with ttk
gitk uses ${NS} to select between the original Tk widgets, and the newer
themed widgets, the former are in the anonymous global namespace, while
the themed ones are in the ttk namespace.  As gitk uses only themed
widgets, this indirection now serves no purpose. Let's switch to
explicit use of ttk:: via global search/replace. More simplification,
including removal of the NS variable, is kept for a later patch to keep
this one smaller.

Signed-off-by: Mark Levedahl <mlevedahl@gmail.com>